Add some tests, set up mocking
