--- OLD_sec_edgar_downloader/_version.py
@@ -1,1 +1,0 @@
-__version__ = "4.3.0"
