--- OLD_tests/sample-filings/aapl-4-20201013.xml
@@ -1,93 +1,0 @@
-<?xml version="1.0"?>
-<ownershipDocument>
-
-    <schemaVersion>X0306</schemaVersion>
-
-    <documentType>4</documentType>
-
-    <periodOfReport>2020-10-09</periodOfReport>
-
-    <notSubjectToSection16>0</notSubjectToSection16>
-
-    <issuer>
-        <issuerCik>0000320193</issuerCik>
-        <issuerName>Apple Inc.</issuerName>
-        <issuerTradingSymbol>AAPL</issuerTradingSymbol>
-    </issuer>
-
-    <reportingOwner>
-        <reportingOwnerId>
-            <rptOwnerCik>0001513362</rptOwnerCik>
-            <rptOwnerName>Maestri Luca</rptOwnerName>
-        </reportingOwnerId>
-        <reportingOwnerAddress>
-            <rptOwnerStreet1>ONE APPLE PARK WAY</rptOwnerStreet1>
-            <rptOwnerStreet2></rptOwnerStreet2>
-            <rptOwnerCity>CUPERTINO</rptOwnerCity>
-            <rptOwnerState>CA</rptOwnerState>
-            <rptOwnerZipCode>95014</rptOwnerZipCode>
-            <rptOwnerStateDescription></rptOwnerStateDescription>
-        </reportingOwnerAddress>
-        <reportingOwnerRelationship>
-            <isDirector>0</isDirector>
-            <isOfficer>1</isOfficer>
-            <isTenPercentOwner>0</isTenPercentOwner>
-            <isOther>0</isOther>
-            <officerTitle>Senior Vice President, CFO</officerTitle>
-            <otherText></otherText>
-        </reportingOwnerRelationship>
-    </reportingOwner>
-
-    <nonDerivativeTable>
-        <nonDerivativeTransaction>
-            <securityTitle>
-                <value>Common Stock</value>
-                <footnoteId id="F1"/>
-            </securityTitle>
-            <transactionDate>
-                <value>2020-10-09</value>
-            </transactionDate>
-            <transactionCoding>
-                <transactionFormType>4</transactionFormType>
-                <transactionCode>S</transactionCode>
-                <equitySwapInvolved>0</equitySwapInvolved>
-            </transactionCoding>
-            <transactionAmounts>
-                <transactionShares>
-                    <value>243431</value>
-                </transactionShares>
-                <transactionPricePerShare>
-                    <value>116.89</value>
-                    <footnoteId id="F2"/>
-                </transactionPricePerShare>
-                <transactionAcquiredDisposedCode>
-                    <value>D</value>
-                </transactionAcquiredDisposedCode>
-            </transactionAmounts>
-            <postTransactionAmounts>
-                <sharesOwnedFollowingTransaction>
-                    <value>110272</value>
-                </sharesOwnedFollowingTransaction>
-            </postTransactionAmounts>
-            <ownershipNature>
-                <directOrIndirectOwnership>
-                    <value>D</value>
-                </directOrIndirectOwnership>
-            </ownershipNature>
-        </nonDerivativeTransaction>
-    </nonDerivativeTable>
-
-    <derivativeTable></derivativeTable>
-
-    <footnotes>
-        <footnote id="F1">This transaction was made pursuant to a Rule 10b5-1 trading plan adopted by the reporting person on May 13, 2019.</footnote>
-        <footnote id="F2">This transaction was executed in multiple trades at prices ranging from $116.81 to $116.95; the price reported above reflects the weighted average sale price. The reporting person hereby undertakes to provide full information regarding the number of shares and prices at which the transactions were effected upon request to the SEC staff, Apple, or a security holder of Apple.</footnote>
-    </footnotes>
-
-    <remarks></remarks>
-
-    <ownerSignature>
-        <signatureName>/s/ Sam Whittington, Attorney-in-Fact for Luca Maestri</signatureName>
-        <signatureDate>2020-10-13</signatureDate>
-    </ownerSignature>
-</ownershipDocument>
