+++ tests/conftest.py
@@ -1,5 +1,6 @@
 """Pytest fixtures for testing suite."""
 import shutil
+from unittest.mock import patch
 
 import pytest
 
@@ -11,6 +12,11 @@ EMAIL = "foo@bar.com"
 
 @pytest.fixture(scope="function")
 def downloader(tmp_path):
-    dl = Downloader(COMPANY_NAME, EMAIL, tmp_path)
-    yield dl, tmp_path
-    shutil.rmtree(tmp_path)
+    with patch(
+        "sec_edgar_downloader.Downloader.get_ticker_to_cik_mapping"
+    ) as mock_ticker_cik_mapping:
+        # Mock ticker to CIK mapping to prevent a call to SEC API
+        mock_ticker_cik_mapping.return_value = {"AAPL": "0000320193"}
+        dl = Downloader(COMPANY_NAME, EMAIL, tmp_path)
+        yield dl, tmp_path
+        shutil.rmtree(tmp_path)
